// shared/entities/company-settings.entity.d.ts
export declare class CompanySettings {
    id: string;
    name: string;
    address: string;
    city: string;
    postalCode: string;
    country: string;
    ico: string;
    dic: string;
    phone: string;
    email: string;
    website: string;
    logoUrl: string;
    notes: string;
    createdAt: Date;
    updatedAt: Date;
}